classes now populate to calendar
Classes populate to calendar, there is a momentary pop in currently. Classes are called using an AJAX call which can be found in home_owner. This references get_classes in views.py. Which assembles an array of JSON objects, which the ajax call retrieves and then translates for the calendar to then populate.

Do to: Calendar cuts off, unable to fully see days in the weekend. Classes are not clickable into yet.